[f1] fix int8 overflow in cell-type transfer bucket key computation
Cast collapsed grid to int64 before multiplying by 1000.

+++ tasks/astar/src/astar/student/predictor/cell_type_transfer.py
@@ -319,7 +319,7 @@ class CellTypeTransferPredictor(BaseRoundPredictor):
             fn_capped = np.minimum(forest_neighbors, 8)
 
             keys = (
-                collapsed * 1000
+                collapsed.astype(np.int64) * 1000
                 + prox_buckets * 100
                 + coast_flags * 10
                 + fn_capped
